Create endpoint POST and methods to remove staff

+++ backend/src/main/java/mg/healthpoint/service/StaffService.java
@@ -111,9 +111,11 @@ public class StaffService {
 		
 	}
 	
-	public void delete(Staff staff) {
+	public void delete(Authentication auth, Staff staff) throws NotFoundException, ForbiddenException {
 		
+		this.checkAdmin(auth);
 		staffRepository.delete(staff);
+		accountService.remove(staff.getAccount());
 		
 	}
